Weebreader: Rename classes
